Implemented JAR minimization
This commit adds two new configuration options to the 'shadowJar' task.
When 'minimize' is set the shadow JAR would only include the classes
for the project the task operates on and its dependencies. The user can
protect other classes from minimization by exluding their module.

Here's an example configuration

    shadowJar {
      minimize {
        exclude(dependency('junit:junit:.*'))
      }
    }

N.B. imported but unused classes are not considered as dependencies.

Change-Id: Ibbc2bdb2b532ab294883b109cae17cc17e4b0e28